Added some styling changes
